mudanças leves , deixando o codigo mais enchuto , logcico não foi mudada

+++ ticket_manager/ticket_manager/routers/clients.py
@@ -1,10 +1,11 @@
 
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import or_, select
 
 from ticket_manager.database import session_db
 from ticket_manager.models import Client
-from ticket_manager.schema import ClientList, ClientPublicShcema, ClientSchema
+from ticket_manager.schema import ClientList, ClientPublic, ClientSchema
+from ticket_manager.security import login_required
 from ticket_manager.services.clients_services import (
     get_client_by_id,
     parse_client,
@@ -17,9 +18,13 @@ clients_router = APIRouter(prefix='/clients', tags=['clients'])
 @clients_router.post(
     "/",
     status_code=201,
-    response_model=ClientPublicShcema
+    response_model=ClientPublic
     )
-def create_client(client: ClientSchema, session: session_db):
+def create_client(
+    client: ClientSchema,
+    session: session_db,
+    loged_user=Depends(login_required)
+    ):
     new_client = Client(
         name=client.name,
         company_name=client.company_name or "Não cadastrado",
@@ -29,20 +34,21 @@ def create_client(client: ClientSchema, session: session_db):
     session.commit()
     session.refresh(new_client)
 
-    new_parse_client = parse_client_public(new_client)
-
-    return new_parse_client
+    return parse_client_public(new_client)
 
 
 @clients_router.get(
     '/{user_id}',
     status_code=200,
-    response_model=ClientPublicShcema,
+    response_model=ClientPublic
     )
-def get_client(client_id: int, session: session_db):
+def get_client(
+    client_id: int,
+    session: session_db,
+    loged_user=Depends(login_required)
+    ):
     client_query = get_client_by_id(session, client_id)
-    parse_client = parse_client_public(client_query)
-    return parse_client
+    return parse_client_public(client_query)
 
 
 @clients_router.put(
@@ -54,8 +60,15 @@ def update_client(
     client_id: int,
     client: ClientSchema,
     session: session_db,
+    loged_user=Depends(login_required)
 ):
-    existing_client = session.get(Client, client_id)
+    existing_client = get_client_by_id(session, client_id)
+
+    if not any([client.name, client.company_name, client.phone]):
+        raise HTTPException(
+            status_code=400,
+            detail="At least one field must be provided for update"
+            )
 
     existing_client.name = client.name or existing_client.name
     existing_client.company_name = (
@@ -66,9 +79,7 @@ def update_client(
     session.commit()
     session.refresh(existing_client)
 
-    updated_parse_client = parse_client(existing_client)
-
-    return updated_parse_client
+    return parse_client(existing_client)
 
 
 @clients_router.get(
@@ -76,7 +87,11 @@ def update_client(
     status_code=200,
     response_model=ClientList
     )
-def search_clients(session: session_db, search_term: str | None = None):
+def search_clients(
+    session: session_db,
+    search_term: str | None = None,
+    loged_user=Depends(login_required)
+    ):
 
     if search_term:
         query = select(Client).where(
@@ -87,23 +102,31 @@ def search_clients(session: session_db, search_term: str | None = None):
         )
         clients = session.scalars(query).all()
         if not clients:
-            raise HTTPException(status_code=404,
-                                detail="Nenhum cliente encontrado")
-
+            raise HTTPException(
+                status_code=404,
+                detail="Nenhum cliente encontrado"
+                )
         client_list = [
-            ClientPublicShcema.model_validate(client) for client in clients
+            ClientPublic.model_validate(client) for client in clients
             ]
         return {'clientlist': client_list}
     else:
         clients = session.query(Client).all()
         client_list = [
-            ClientPublicShcema.model_validate(client) for client in clients
+            ClientPublic.model_validate(client) for client in clients
             ]
         return {'clientlist': client_list}
 
 
 @clients_router.delete('/{client_id}', status_code=204)
-def delete_client(session: session_db, client_id: int):
+def delete_client(
+    session: session_db,
+    client_id: int,
+    loged_user=Depends(login_required),
+    ):
     client = session.get(Client, client_id)
+    if not client:
+        raise HTTPException(status_code=404, detail="Client not found")
+
     session.delete(client)
     session.commit()
